Clear up DefaultFileHierarchySet
